As DB is stored on persistence storage, don't disclose DB
The application’s _protected_ method (see: gabber/views.py) sends items
from the protected folder (based on a lookup of the given name). Given
the database is stored their (and its entirety) and it will be in an
active, public-facing deployment (rather than lab deployment where the
database being public is of no concern), then access to this will be
denied.

Of course, haxzors need to know the name of the database (similar to a
private key), so now even with that knowledge they will be denied the
ease of downloading such information.

A better solution is to store it above the root, which is not possible
with dokku. Instead, using another provided besides dokku (such as AWL
with Postgres) would be simpler and safer, and is what will be
undertaken in the future.