Store complex needs if they are provided
This is for many participants in an interview, and now stores the
interview ID correctly.

_Note:_ it is not certain that this way to acquire the interview ID is
suitable as if two requests are made at the _Exact_ same time, then,
argh … bad times.

An alternative way would be to add the interview before the if
statement and flush (thereby inserting only an interview without any
relationships), then updating the relationships after as this would
allow us to access the ID of that interview, rather than attempting to
infer it.