Add a customized anonymous user
This is required to create custom access control throughout the
website, such as showing the edit/delete options to admins and not an
anonymous user. In short, this allows us to call
`current_user.whatever()`, even with an anonymous user as a result
(probably None) would be returned.

This also helps us track anonymous users on the website.