:wrench: #1 Upload project image to S3
All project images (and future images, e.g. profile pics) are stored in a static folder that is public; each file is a named a random UUID.

One current issue is that we choose a default image if we had issue uploading the file. That way, the project is created (which the user wanted) and they can update the project image later, however, there is currently no response to the user for this as it's quite an edge case to be returning errors.